Added init modules to namespaces. Reformated structure. Renamed namespaces. Refactored setup.
