+++ .pre-commit-config.yaml
@@ -1,5 +1,8 @@
 ﻿# See https://pre-commit.com for more information
 # See https://pre-commit.com/hooks.html for more hooks
+# pre-commit sample-config | out-file .pre-commit-config.yaml -encoding utf8
+# pre-commit install
+# pre-commit run --all-files
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
   rev: v4.0.1
