+++ src/filter_core/FilterABC.py
