+++ src/image_filter/MeanFilter.py
@@ -1,10 +1,10 @@
 import numpy as np
-from Core.FilterBase import FilterBase
+from filter_core.FilterBase import FilterBase
 
 
-class MedianFilter(FilterBase):
+class MeanFilter(FilterBase):
     def __init__(self, img, kernel: int, stride: int, padding: int) -> None:
-        super().__init__(img, kernel, stride, padding)
+        super().__init__(img=img, kernel=kernel, stride=stride, padding=padding)
 
         self._execute()
         self._convert_to_numpy_array()
@@ -17,4 +17,4 @@ class MedianFilter(FilterBase):
                 temp = self.img[i : i + self.kernel, j : j + self.kernel]
                 if temp.size == self.kernel ** 2:
                     # O(1)
-                    self.smoothCriminal.append(np.median(temp))
+                    self.smoothCriminal.append(np.mean(temp))
