+++ src/image_filter/MedianFilter.py
@@ -1,10 +1,10 @@
 import numpy as np
-from Core.FilterBase import FilterBase
+from filter_core.FilterBase import FilterBase
 
 
-class MeanFilter(FilterBase):
+class MedianFilter(FilterBase):
     def __init__(self, img, kernel: int, stride: int, padding: int) -> None:
-        super().__init__(img, padding)
+        super().__init__(img, kernel, stride, padding)
 
         self._execute()
         self._convert_to_numpy_array()
@@ -17,4 +17,4 @@ class MeanFilter(FilterBase):
                 temp = self.img[i : i + self.kernel, j : j + self.kernel]
                 if temp.size == self.kernel ** 2:
                     # O(1)
-                    self.smoothCriminal.append(np.mean(temp))
+                    self.smoothCriminal.append(np.median(temp))
