+++ src/tensor_filter.egg-info/PKG-INFO
@@ -1,0 +1,10 @@
+Metadata-Version: 1.0
+Name: tensor-filter
+Version: 0.0.1
+Summary: Python package that applies filters to images
+Home-page: UNKNOWN
+Author: Joel Castro
+Author-email: UNKNOWN
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
