+++ src/tensor_filter.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+README
+setup.py
+src/filter_core/FilterABC.py
+src/filter_core/FilterBase.py
+src/filter_core/__init__.py
+src/image_filter/MeanFilter.py
+src/image_filter/MedianFilter.py
+src/image_filter/__init__.py
+src/tensor_filter.egg-info/PKG-INFO
+src/tensor_filter.egg-info/SOURCES.txt
+src/tensor_filter.egg-info/dependency_links.txt
+src/tensor_filter.egg-info/requires.txt
+src/tensor_filter.egg-info/top_level.txt
