+++ src/tensor_filter.egg-info/requires.txt
@@ -1,0 +1,1 @@
+numpy
