+++ src/tensor_filter.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+filter_core
+image_filter
