+++ tests/mean_filter_test.py
@@ -1,5 +1,5 @@
 import numpy as np
-from ImageFilter import MeanFilter
+from image_filter.MeanFilter import MeanFilter
 
 img = np.random.rand(768, 1024)
 av_filter = MeanFilter(img, kernel=4, stride=4, padding=0)
